Quality fixes (#137)
* Code quality cleanup
* Update UTF-16 code to match RFC 3629 update to UTF-8
* Pop and ignore PC
* Always use -Werror

+++ Emulator/Screen/TFLScreenManager.cpp
@@ -464,7 +464,7 @@ public:
     }
 
 	void newRGBSize(int w, int h) {
-	    size_t sz = w * h;
+	    size_t sz = (size_t) w * (size_t) h;
 		if (sz != (size_t)(rgbWidth_ * rgbHeight_)) {
 			if (rgbData_) 
 				::free(rgbData_);
